Call to mkldnn_matmul from aten::addmm on AArch64
We have noticed that on BERT_pytorch in torchbenchmark
majority of time is spent in running GEMM in aten:addmm.
At the moment this calls into BLAS routine, but on AArch64
it will be faster if it calls into mkldnn_matmul. Performance
wise compared to build with OpenBLAS it runs faster 1.2x
faster on 16 cores with batch size of 8 on Graviton3,
while if fast math mode (mkldnn_matmul exposes through oneDNN
and Arm Compute Library option to run GEMM with FP32 inputs using
BBF16 operations) is enabled then it is 2.3x